backend/requirements.txt
flask
google-cloud-bigtable
google-cloud-spanner
redis
firebase-admin
gunicorn
flask-cors
